Backend/requirements.txt
aenum==3.1.15
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==23.2.0
boto3==1.34.136
botocore==1.34.136
certifi==2024.6.2
charset-normalizer==3.3.2
Deprecated==1.2.14
Django==5.0.6
djangorestframework==3.15.2
frozenlist==1.4.1
future==1.0.0
idna==3.7
jmespath==1.0.1
line-bot-sdk==3.11.0
multidict==6.0.5
pillow==10.3.0
pydantic==2.7.4
pydantic_core==2.18.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.31.0
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.2
wrapt==1.16.0
yarl==1.9.4